Fix for bug identified in github issue #16.
Error in the code showed that a task was being assigned to a workcenter
during constrained planning only if it was met on time; if late, then
the workcenter assignment step was skipped.

Fixing that bug revealed another bug in the search order for alternate
workcenters; a wrong check (lesser than vs. lesser than or equal to) was
resulting in a delayed TaskPlan being assigned to the alternate resource
instead of the primary resource if the TaskPlan was delayed on both
resources.

Both problems have been fixed.  Along the way, some debugging code was
added to Demand.java.