Fixes for github issues #4, #5, and #20.  Details below:
Issue #4 (Database Schema and support):
--------------------------------------
(1) A new folder called "db" has been added with the file
"jobshop_schema.sql" -> this is a sample schema file for a sqlite3
database.

(2) Changes made to the JobShop.java class to support an option file
that allows the user to specify the database connection details; the
loading function will now load from the appropriate database as
specified in the options file.

(3) Output of TaskPlans will be written to the taskplan table in the
database.  Since all of the information is available in the database, no
other output is written to the screen.  The user can write suitable
queries of their choosing to analyze the plan results.

Issue #5 (Improved regression testing support):
---------------------------------------------
(1) Added the tests/expect subdirectory with expected output files.

(2) Makefile modified to add an option to run tests ("make tests").
This will run all of the tests, one by one and compare the expect files
in the above expect directory against the output file and print a
pass/fail message.

(3) A single testplan can be run using "make TESTNAME=testXXXX tests" to
run a single test in the directory tests/testXXXX

Issue #20 (Copyright information):
--------------------------------
Appropriate notice added to all java source files as specified in the
issue.

Miscellaneous (related to above issues tangentially):
---------------------------------------------------
(1) Command line arguments changed; now a single options file will be
accepted as argument. Refer to wiki for details

(2) Flatfiles for all testcases renamed where applicable to match the
corresponding database table name (and header line also modified
according to schema).  Some data changes to match the schema.  For
example, instead of a unique taskid, we make the combination of taskid +
sku to be unique.