Adding utility scripts to load a sqlite database.
This fix addresses github issue #21.  As a result of developing suitable
scripts for this issue, we have also had to modify some of the data
files (order of the data columns) of the test plans and the JobShop
class which loads tests plans.

Two methods are provided for loading datasets into the database:

(1) using the native sqlite loader; 2 new SQL files are provided to
create staging tables into which data can be loaded and then moved over
to the actual tables.  A shell script is provided to coordinate all of
this.

(2) A python script is provided that does all of the work.

Please refer to the wiki which will have more details on how to use
these scripts.