Adding documentation files for the test cases - github issue #22
html files have been added for the test cases; along the way, minor
changes made to test0011 and test0012 to "build a story" of how
the solver works as the data changes and more complex combinations
of supported features are used.

A small bug in the run_tests.sh script fixed that caused an error
when trying to run a single test plan.